requirements.txt
streamlit
litellm
requests
beautifulsoup4
scikit-learn
numpy
pandas
pyvis
python-dotenv
aiohttp
wordcloud
matplotlib
